refactor: move textArea code to separate class
